better handling of vars vs. functions in responses
If a student writes '3 sin x', the parser identified sin as the sine function. This was done in order to throw a syntax error (warning the student) due to the missing parens.

Now, we consider sin as a variable (because there are no parens). This will still lead to an error for numeric answer types (number, numeric, numerical formula), but not for algebraic formulas. Hence, some code was added to disallow variables named like functions in student answers. This is still backwards compatible, because in earlier versions, variables could not have function names, even on the teacher side.