improve robustness when importing/restoring bad data (#273)
Importing questions or restoring them from a backup could fail, if options specific to the Formulas question type (like definition of random or global variables) or question parts were missing, e. g. due to corruption of the DB. With this fix, the process should still be able to finish, even if some of the data is bad. Appropriate notices or warnings will be output. If options are missing, defaults will be used. For missing parts, it is obviously not possible to do anything. Those questions will simply not work, until they are edited and manually repaired.