allow spaces around variable name in MC input (#292)
Although the documented syntax never uses spaces around the colon in the definition of a multiple-choice input, legacy versions did not actually prohibit it, so {_0: foo :MCE} was an undocumented use, but it worked. This PR makes sure that existing questions containing such rogue spaces will continue to work properly, as such spaces might especially be introduced on Moodle instances that use the multilang filter.